split actors page into two tabs
