+++ middleware/routes.global.ts
@@ -19,5 +19,9 @@ export default defineNuxtRouteMiddleware((to, from) => {
     if (to.path === "/config/") {
       return navigateTo("/config/governance/");
     }
+
+    if (to.path === "/actors/") {
+      return navigateTo("/actors/protocol/");
+    }
   }
 });
